Add size display and delete confirmation

+++ ui/explorer.py
@@ -2,7 +2,7 @@ import itertools
 import os.path
 from math import floor
 from tkinter import *
-from tkinter import filedialog
+from tkinter import filedialog, messagebox
 
 from adb_device import AdbDevice
 from config import *
@@ -16,22 +16,22 @@ def explorer(device_name, root: Tk):
     root.minsize(*APP_CONFIG['min_size'])
     root.maxsize(*APP_CONFIG['max_size'])
     root.title(f'{device.name} - {APP_CONFIG["app_name"]}')
+    root.geometry('800x600')
 
-    head_frame = LabelFrame(root, text='Header', padx=10, pady=10)
-    head_frame.pack(fill=X, anchor=N)
-    foot_frame = LabelFrame(root, text='Footer', padx=10, pady=10)
-    foot_frame.pack(fill=X, anchor=S, side=BOTTOM)
+    header_frame = LabelFrame(root, text='Header', padx=10, pady=10)
+    header_frame.pack(fill=X, anchor=N)
+    footer_frame = LabelFrame(root, text='Footer')
+    footer_frame.pack(fill=X, anchor=S, side=BOTTOM)
 
-    address_bar_text = StringVar(head_frame, device.current_dir)
-    bottom_bar_text = StringVar(head_frame)
-    address_bar = Entry(head_frame, textvariable=address_bar_text, borderwidth=3)
+    address_bar_text = StringVar(header_frame, device.current_dir)
+    bottom_bar_text = StringVar(header_frame)
+    address_bar = Entry(header_frame, textvariable=address_bar_text, borderwidth=3)
     address_bar.pack(fill=X, expand=1, side=LEFT)
-    bottom_bar = Label(foot_frame, textvariable=bottom_bar_text, borderwidth=10)
+    bottom_bar = Label(footer_frame, textvariable=bottom_bar_text, borderwidth=10)
     bottom_bar.grid(row=1, column=0, columnspan=4, sticky=W)
 
-    view_frame = LabelFrame(root, text='Files', padx=10, pady=10)
-    view_frame.pack_propagate(False)
-    view_frame.pack(fill=BOTH, expand=1)
+    base_view_frame = LabelFrame(root, text='Base View Frame', padx=10, pady=10)
+    base_view_frame.pack(fill=BOTH, expand=1)
     view_items = []
 
     def file_ui_position_generator():
@@ -46,38 +46,37 @@ def explorer(device_name, root: Tk):
             device.pull_file(file.get_full_path(), target)
 
     def delete_file(obj: File):
-        device.delete_file(obj)
-        enter_path()
+        result = messagebox.askokcancel('Confirm Deleting', f'{obj.name} will be deleted')
+        if result:
+            device.delete_file(obj)
+            enter_path()
 
     def get_file_ui_element(file: File):
-        file_frame = LabelFrame(view_frame, text=file.get_type_mark(), padx=3, pady=3, width=FILE_ITEM_CONFIG['width'],
-                                height=FILE_ITEM_CONFIG['height'])
-        file_frame.pack_propagate(False)
+        file_item_frame = LabelFrame(base_view_frame, text=file.get_type(), padx=3, pady=3,
+                                     width=FILE_ITEM_CONFIG['width'],
+                                     height=FILE_ITEM_CONFIG['height'])
+        file_item_frame.pack_propagate(False)
 
-        file_name = Label(file_frame, text=file.name, wraplength=FILE_ITEM_CONFIG['width'] - 100)
-        file_size = Label(file_frame, text=str(file.size))
-        file_name.pack(side=LEFT)
+        if len(file.name) > FILE_ITEM_CONFIG['max_display_name_length']:
+            f_name = file.name[0:FILE_ITEM_CONFIG['max_display_name_length'] + 1] + '...'
+        else:
+            f_name = file.name
+
+        file_name = Label(file_item_frame, text=f_name, wraplength=FILE_ITEM_CONFIG['width'] - 10, anchor=NW)
+        file_name.pack(fill=X)
+        file_size = Label(file_item_frame, text=file.get_readable_size())
         file_size.pack(side=LEFT)
 
-        inner_frame = LabelFrame(file_frame, padx=3, pady=3)
-        inner_frame.pack(side=RIGHT)
-
-        button_download = Button(inner_frame, text='↓', command=lambda: download_file(file))
-        button_delete = Button(inner_frame, text='X', command=lambda: delete_file(file))
-
-        if file.is_file:
-            button_enter_state = DISABLED
-        else:
-            button_enter_state = NORMAL
-
-        button_enter = Button(inner_frame, text='↗', command=lambda: enter_path_with_obj(file),
-                              state=button_enter_state)
+        button_download = Button(file_item_frame, text='↓', command=lambda: download_file(file))
+        button_delete = Button(file_item_frame, text='X', command=lambda: delete_file(file))
 
         button_download.pack(side=RIGHT)
         button_delete.pack(side=RIGHT)
-        button_enter.pack(side=RIGHT)
+        if not file.is_file:
+            button_enter = Button(file_item_frame, text='↗', command=lambda: enter_path_with_obj(file))
+            button_enter.pack(side=RIGHT)
 
-        return file_frame
+        return file_item_frame
 
     def enter_path():
         file_position_generator = file_ui_position_generator()
@@ -120,9 +119,9 @@ def explorer(device_name, root: Tk):
             print(f'Folder push succeed:{folder_to_upload}')
         enter_path()
 
-    button_enter_path = Button(head_frame, text='→/⚪', command=enter_path)
-    button_upload_file = Button(head_frame, text='↑ File(s)', command=upload_files)
-    button_upload_folder = Button(head_frame, text='↑ Folder', command=upload_folder)
+    button_enter_path = Button(header_frame, text='→/⚪', command=enter_path)
+    button_upload_file = Button(header_frame, text='↑ File(s)', command=upload_files)
+    button_upload_folder = Button(header_frame, text='↑ Folder', command=upload_folder)
     button_upload_folder.pack(side=RIGHT)
     button_upload_file.pack(side=RIGHT)
     button_enter_path.pack(side=RIGHT)
